new: adjust cuGraph to ArangoDB for increased accessibility to python-arango interface (#18)
* #17: initial commit

* fix: typo

* Update build.yml

* disable overwrite

(the self-hosted runners are using the same arangodb instance)

* attempt: re-enable overwrite

* Update README.md

* Update README.md

* remove: __validate_attributes()

Uncessary noise, replaced in favor of proper docstring examples

* chg: #18 (after redefining scope)

* @aMahanna new: CodeQL Action v2

* fix: mypy

* fix: typo

* bump: python-arango version

* cleanup: documentation

+++ adbcug_adapter/adapter.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 import logging
 from collections import defaultdict
-from typing import Any, DefaultDict, Dict, List, Set, Tuple, Union
+from typing import Any, DefaultDict, Dict, List, Optional, Set, Tuple, Union
 
 from arango.cursor import Cursor
 from arango.database import Database
@@ -84,14 +84,26 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
             will set the edge weight value to 0.
         :type edge_attr: str
         :param query_options: Keyword arguments to specify AQL query options when
-            fetching documents from the ArangoDB instance.
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A Multi-Directed cuGraph Graph.
         :rtype: cugraph.structure.graph_classes.MultiDiGraph
         :raise ValueError: If missing required keys in metagraph
+
+        Here is an example entry for parameter **metagraph**:
+
+        .. code-block:: python
+        {
+            "vertexCollections": {
+                "account": {}, # cuGraph does not support node attributes
+            },
+            "edgeCollections": {
+                "transaction": {}, # cuGraph does not support edge attributes
+            },
+        }
         """
         logger.debug(f"Starting arangodb_to_cugraph({name}, ...):")
-        self.__validate_attributes("graph", set(metagraph), self.METAGRAPH_ATRIBS)
 
         # Maps ArangoDB vertex IDs to cuGraph node IDs
         adb_map: Dict[str, Dict[str, Union[CUGId, str]]] = dict()
@@ -147,7 +159,8 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
             will set the edge weight value to 0.
         :type edge_attr: str
         :param query_options: Keyword arguments to specify AQL query options when
-            fetching documents from the ArangoDB instance.
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A Multi-Directed cuGraph Graph.
         :rtype: cugraph.structure.graph_classes.MultiDiGraph
@@ -170,7 +183,8 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
             will set the edge weight value to 0.
         :type edge_attr: str
         :param query_options: Keyword arguments to specify AQL query options when
-            fetching documents from the ArangoDB instance.
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A Multi-Directed cuGraph Graph.
         :rtype: cugraph.structure.graph_classes.MultiDiGraph
@@ -187,11 +201,12 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
         self,
         name: str,
         cug_graph: CUGGraph,
-        edge_definitions: List[Json],
-        batch_size: int = 1000,
+        edge_definitions: Optional[List[Json]] = None,
         keyify_nodes: bool = False,
         keyify_edges: bool = False,
+        overwrite_graph: bool = False,
         edge_attr: str = "weight",
+        **import_options: Dict[str, Any],
     ) -> ADBGraph:
         """Create an ArangoDB graph from a cuGraph graph, and a set of edge
         definitions.
@@ -200,13 +215,11 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
         :type name: str
         :param cug_graph: The existing cuGraph graph.
         :type cug_graph: cugraph.classes.graph.Graph
-        :param edge_definitions: List of edge definitions, where each edge definition
-            entry is a dictionary with fields "edge_collection",
-            "from_vertex_collections" and "to_vertex_collections"
-            (see below for example).
-        :type edge_definitions: List[adbcug_adapter.typings.Json]
-        :param batch_size: The maximum number of documents to insert at once
-        :type batch_size: int
+        :param edge_definitions: List of edge definitions, where each edge
+            definition entry is a dictionary with fields "edge_collection",
+            "from_vertex_collections" and "to_vertex_collections" (see below
+            for example). Can be omitted if the graph already exists.
+        :type edge_definitions: List[adbnx_adapter.typings.Json]
         :param keyify_nodes: If set to True, will create custom node keys based on the
             behavior of ADBCUG_Controller._keyify_cugraph_node().
             Otherwise, ArangoDB _key values for vertices will range from 1 to N,
@@ -221,6 +234,13 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
             the edge attribute name used to represent your cuGraph edge weight values
             once transferred into ArangoDB. Defaults to 'weight'.
         :type edge_attr: str
+        :param overwrite_graph: Overwrites the graph if it already exists.
+            Does not drop associated collections.
+        :type overwrite_graph: bool
+        :param import_options: Keyword arguments to specify additional
+            parameters for ArangoDB document insertion. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.collection.Collection.import_bulk
+        :type import_options: Any
         :return: The ArangoDB Graph API wrapper.
         :rtype: arango.graph.Graph
 
@@ -236,16 +256,18 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
         ]
         """
         logger.debug(f"Starting cugraph_to_arangodb('{name}', ...):")
-        for e_d in edge_definitions:
-            self.__validate_attributes(
-                "Edge Definitions", set(e_d), self.EDGE_DEFINITION_ATRIBS
-            )
 
-        self.__db.delete_graph(name, ignore_missing=True)
-        adb_graph: ADBGraph = self.__db.create_graph(name, edge_definitions)
+        if overwrite_graph:
+            logger.debug("Overwrite graph flag is True. Deleting old graph.")
+            self.__db.delete_graph(name, ignore_missing=True)
+
+        if self.__db.has_graph(name):
+            adb_graph = self.__db.graph(name)
+        else:
+            adb_graph = self.__db.create_graph(name, edge_definitions)
 
         adb_v_cols = adb_graph.vertex_collections()
-        adb_e_cols = [e_d["edge_collection"] for e_d in edge_definitions]
+        adb_e_cols = [e_d["edge_collection"] for e_d in adb_graph.edge_definitions()]
 
         has_one_vcol = len(adb_v_cols) == 1
         has_one_ecol = len(adb_e_cols) == 1
@@ -277,8 +299,7 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
                 "adb_key": key,
             }
 
-            adb_vertex = {"_id": adb_v_id}
-            self.__insert_adb_docs(col, adb_documents[col], adb_vertex, batch_size)
+            adb_documents[col].append({"_id": adb_v_id})
 
         from_node_id: CUGId
         to_node_id: CUGId
@@ -309,39 +330,15 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
             if cug_graph.is_weighted():
                 adb_edge[edge_attr] = weight[0]
 
-            self.__insert_adb_docs(
-                col,
-                adb_documents[col],
-                adb_edge,
-                batch_size,
-            )
+            adb_documents[col].append(adb_edge)
 
-        for col, doc_list in adb_documents.items():  # insert remaining documents
-            if doc_list:
-                logger.debug(f"Inserting last {len(doc_list)} documents into '{col}'")
-                self.__db.collection(col).import_bulk(doc_list, on_duplicate="replace")
+        for col, doc_list in adb_documents.items():  # import documents into ArangoDB
+            logger.debug(f"Inserting {len(doc_list)} documents into '{col}'")
+            result = self.__db.collection(col).import_bulk(doc_list, **import_options)
+            logger.debug(result)
 
         logger.info(f"Created ArangoDB '{name}' Graph")
         return adb_graph
-
-    def __validate_attributes(
-        self, type: str, attributes: Set[str], valid_attributes: Set[str]
-    ) -> None:
-        """Validates that a set of attributes includes the required valid
-        attributes.
-
-        :param type: The context of the attribute validation
-            (e.g connection attributes, graph attributes, etc).
-        :type type: str
-        :param attributes: The provided attributes, possibly invalid.
-        :type attributes: Set[str]
-        :param valid_attributes: The valid attributes.
-        :type valid_attributes: Set[str]
-        :raise ValueError: If **valid_attributes** is not a subset of **attributes**
-        """
-        if valid_attributes.issubset(attributes) is False:
-            missing_attributes = valid_attributes - attributes
-            raise ValueError(f"Missing {type} attributes: {missing_attributes}")
 
     def __fetch_adb_docs(self, col: str, query_options: Any) -> Result[Cursor]:
         """Fetches ArangoDB documents within a collection.
@@ -360,29 +357,3 @@ class ADBCUG_Adapter(Abstract_ADBCUG_Adapter):
         """
 
         return self.__db.aql.execute(aql, **query_options)
-
-    def __insert_adb_docs(
-        self,
-        col: str,
-        col_docs: List[Json],
-        doc: Json,
-        batch_size: int,
-    ) -> None:
-        """Insert an ArangoDB document into a list. If the list exceeds
-        batch_size documents, insert into the ArangoDB collection.
-
-        :param col: The collection name
-        :type col: str
-        :param col_docs: The existing documents data belonging to the collection.
-        :type col_docs: List[adbnx_adapter.typings.Json]
-        :param doc: The current document to insert.
-        :type doc: adbnx_adapter.typings.Json
-        :param batch_size: The maximum number of documents to insert at once
-        :type batch_size: int
-        """
-        col_docs.append(doc)
-
-        if len(col_docs) >= batch_size:
-            logger.debug(f"Inserting next {batch_size} batch documents into '{col}'")
-            self.__db.collection(col).import_bulk(col_docs, on_duplicate="replace")
-            col_docs.clear()
